+++ handlers/azure_api_v1/test/get_input.rb
@@ -1,0 +1,15 @@
+{
+  'info' => {
+    'tenent_id' => "",
+    'client_id' => "",
+    'client_secret' => "",
+    'enable_debug_logging' => 'true'
+  },
+  'parameters' => {
+    'error_handling' => 'Error Message',
+    'headers' => '{"ConsistencyLevel": "eventual"}',
+    'method' => 'GET',
+    'path' => '/users?$search="displayName:Chad"',
+    'body' => ''
+  }
+}