+++ changelog.md
@@ -5,7 +5,11 @@ Azure \[bridge-adapters\] (2020-01-28)
 Azure \[bridge-adapters\] (2020-02-07)
   \[kinetic-bridgehub-adapter-azure-rest\] v1.0.2
     * Error handling was preventing token refresh.
-
-Azure \[bridge-adapters\] (2020-02-07)
   \[kinetic-bridgehub-adapter-azure-rest\] v1.0.3
     * The fix for v1.0.2 need additional tweaks.  Error handling for refresh token resolved.
+
+Azure \[handlers\] (2021-05-18)
+  \[azure_api\] v1
+    * Initial commit.
+  \[azure_cosmos_document_crud] v1
+    * Updated readme.